feat: integro cosas de Flor
