+++ client/src/App.jsx
@@ -7,6 +7,7 @@ import FeaturedProducts from "./components/FeaturedProducts";
 import Footer from "./components/Footer";
 import ProductList from "./components/ProductList";
 import ProductDetail from "./components/ProductDetail";
+import ContactoPage from "./components/ContactoPage";
 
 function App() {
   // Estados para los paneles
@@ -16,13 +17,13 @@ function App() {
   // Estado del carrito
   const [cart, setCart] = useState([]);
 
-  // Estados para productos y rutas 
+  // Estados para productos y rutas
   const [products, setProducts] = useState([]);
   const [ruta, setRuta] = useState("/");
   const idBuscado = ruta;
   const indice = products.findIndex((producto) => producto.id === idBuscado);
 
-  // Fetch de productos 
+  // Fetch de productos
   useEffect(() => {
     const url =
       "https://raw.githubusercontent.com/TomasVolpini/ITBA-Muebleria-Jota/refs/heads/develop/server/database/productos.json";
@@ -38,40 +39,37 @@ function App() {
 
   // Función para añadir producto al carrito
   const handleAddToCart = (product) => {
-    setCart(prevCart => {
+    setCart((prevCart) => {
       // Verificar si el producto ya está en el carrito
-      const existingItem = prevCart.find(item => item.id === product.id);
+      const existingItem = prevCart.find((item) => item.id === product.id);
 
       if (existingItem) {
         // Si ya existe, incrementar la cantidad
-        return prevCart.map(item =>
+        return prevCart.map((item) =>
           item.id === product.id
-          ? { ...item, quantity: item.quantity + 1 }
-          : item
+            ? { ...item, quantity: item.quantity + 1 }
+            : item
         );
       }
       // Si no existe, agregarlo con quantity: 1
       return [...prevCart, { ...product, quantity: 1 }];
-  });
-
+    });
   };
 
   // Función para actualizar la cantidad de un producto en el carrito
   const handleUpdateQuantity = (productId, newQuantity) => {
     if (newQuantity < 1) return;
-    
-    setCart(prevCart =>
-      prevCart.map(item =>
-        item.id === productId
-          ? { ...item, quantity: newQuantity }
-          : item
+
+    setCart((prevCart) =>
+      prevCart.map((item) =>
+        item.id === productId ? { ...item, quantity: newQuantity } : item
       )
     );
   };
 
   // Función para eliminar un producto del carrito
   const handleRemoveItem = (productId) => {
-    setCart(prevCart => prevCart.filter(item => item.id !== productId));
+    setCart((prevCart) => prevCart.filter((item) => item.id !== productId));
   };
 
   // Calcular la cantidad total de items en el carrito
@@ -79,44 +77,41 @@ function App() {
 
   return (
     <>
-      <Header 
+      <Header
         cartCount={cartCount}
         onCartOpen={() => setIsCartOpen(true)}
         onMenuToggle={() => setIsMenuOpen(!isMenuOpen)}
         isMenuOpen={isMenuOpen}
+        setRuta={setRuta}
       />
-      
-      <CartPanel 
+
+      <CartPanel
         isOpen={isCartOpen}
         onClose={() => setIsCartOpen(false)}
         cartItems={cart}
         onUpdateQuantity={handleUpdateQuantity}
         onRemoveItem={handleRemoveItem}
       />
-      
+
       {ruta === "/" && (
         <>
           <Hero />
-          <FeaturedProducts 
-            products={products}    
-            setRuta={setRuta}       
-          />
-          <ProductList 
-            products={products} 
-            setRuta={setRuta} 
-          />
+          <FeaturedProducts products={products} setRuta={setRuta} />
+          <ProductList products={products} setRuta={setRuta} />
         </>
       )}
-      
+
       {ruta !== "/" && products[indice] && (
-          <ProductDetail
-            products={products[indice]}
-            setRuta={setRuta}
-            ruta={ruta}
-            onAddToCart={handleAddToCart}
-          />
+        <ProductDetail
+          products={products[indice]}
+          setRuta={setRuta}
+          ruta={ruta}
+          onAddToCart={handleAddToCart}
+        />
       )}
-      
+
+      {ruta === "/contacto" && <ContactoPage></ContactoPage>}
+
       <Footer />
     </>
   );
