+++ client/src/components/BentoGrid.jsx
@@ -1,0 +1,16 @@
+const BentoGrid = ({ children }) => {
+  return (
+    <div
+      style={{
+        display: "grid",
+        gridTemplateColumns: "repeat(auto-fit, minmax(250px, 1fr))",
+        gap: "20px",
+        padding: "20px",
+      }}
+    >
+      {children}
+    </div>
+  );
+};
+
+export default BentoGrid;
