+++ client/src/components/ContactoCard.jsx
@@ -1,0 +1,25 @@
+const ContactoCard = () => {
+  return (
+    <div className="card contacto-card">
+      <h3>Contacto y Créditos</h3>
+      <p>
+        <strong>Showroom</strong>
+      </p>
+      <p>
+        Av. Corrientes 1234
+        <br />
+        Buenos Aires, Argentina
+      </p>
+      <p>
+        <strong>Horarios</strong>
+      </p>
+      <p>
+        Lun - Vie: 9:00 - 18:00
+        <br />
+        Sáb: 10:00 - 14:00
+      </p>
+    </div>
+  );
+};
+
+export default ContactoCard;
