+++ client/src/components/MarcaCard.jsx
@@ -1,0 +1,10 @@
+const MarcaCard = () => {
+  return (
+    <div className="card marca-card">
+      <h3>HJ</h3>
+      <p>ARTESANÍA</p>
+    </div>
+  );
+};
+
+export default MarcaCard;
