+++ client/src/components/TallerCard.jsx
@@ -1,0 +1,9 @@
+const TallerCard = () => {
+  return (
+    <div className="card taller-card">
+      <h3>Nuestro Taller</h3>
+    </div>
+  );
+};
+
+export default TallerCard;
