--- components/contacto/Contacto.css
@@ -1,97 +1,0 @@
-.contacto-container {
-  display: flex;
-  justify-content: center;
-  margin-top: 3rem;
-  padding: 2rem;
-   max-width: 480px;
-}
-
-.grid-contacto {
-  display: flex;
-  gap: 2rem;
-  flex-wrap: wrap;
-  justify-content: center;
-  display: inline-block;
-}
-
-.card-formulario {
-  background: #fff;
-  padding: 2rem;
-  border-radius: 16px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  flex: 1 1 420px;
-}
-
-.card-formulario h2 {
-  color: #5b2e0f;
-  margin-bottom: 1rem;
-}
-
-.card-formulario label {
-  display: block;
-  font-weight: bold;
-  margin-top: 0.8rem;
-}
-
-.card-formulario input,
-.card-formulario textarea {
-  width: 100%;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 0.6rem;
-  margin-top: 0.4rem;
-}
-
-.btn-enviar {
-  background-color: #8a4b2d;
-  color: #fff;
-  border: none;
-  border-radius: 8px;
-  padding: 0.8rem;
-  margin-top: 1rem;
-  width: 100%;
-  cursor: pointer;
-  font-weight: bold;
-}
-
-.mensaje-exito {
-  margin-top: 1rem;
-  background-color: #d7f3e3;
-  color: #2d7a46;
-  padding: 0.6rem;
-  border-radius: 6px;
-  text-align: center;
-  font-weight: 600;
-}
-
-.grid-cards {
-  display: grid;
-  grid-template-columns: repeat(2, 220px);
-  gap: 1.5rem;
-}
-
-.card {
-  background: #fff;
-  padding: 1.5rem;
-  border-radius: 16px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.08);
-  color: #3e2c2c;
-  text-align: left;
-  width: auto;
-  display: inline-block;
-}
-
-.taller-card {
-  background-color: #fcefb0;
-  text-align: center;
-  font-weight: bold;
-  color: #5b2e0f;
-}
-
-.marca-card {
-  background-color: #1c1c1c;
-  color: #fff;
-  text-align: center;
-  font-weight: bold;
-  letter-spacing: 1px;
-}
